Finished function for checking areas for having cropland in them.

+++ trunk/geoprocessing_cpp/probabilityLayer/agreementTable.cpp
@@ -34,7 +34,7 @@ agreementTableT::agreementTableT(const agreementTableT & g)
 	numClasses = g.numClasses;
 	tmpAgreementTable = g.tmpAgreementTable;
 	priorMap = new int[numClasses];
-	for (int i = 0; i < numClasses; i++)
+	for (size_t i = 0; i < numClasses; i++)
 	{
 		priorMap[i] = g.priorMap[i];
 	}
@@ -50,7 +50,7 @@ agreementTableT & agreementTableT::operator = (const agreementTableT & g)
 		numClasses = g.numClasses;
 		tmpAgreementTable = g.tmpAgreementTable;
 		priorMap = new int[numClasses];
-		for (int i = 0; i < numClasses; i++)
+		for (size_t i = 0; i < numClasses; i++)
 		{
 			priorMap[i] = g.priorMap[i];
 		}
